Pull walk() into its own module, add arg to cb()
